Rename dialog class to avoid confusion
