+++ client/src/views/Dashboard.js
@@ -11,7 +11,7 @@ import logo from '../images/everyMealLogo.png';
 import Page from '../components/page/Page';
 import ConsumptionGraph from '../components/charts/Chart';
 import Select from '../components/buttons/SimpleSelect';
-import AddMealButton from '../components/buttons/DialogButton';
+import AddMealDialog from '../components/dialogs/AddMealDialog';
 import ProgressCircle from '../components/charts/ProgessCircle';
 import Legend from '../components/charts/Legend'
 
@@ -58,7 +58,7 @@ export class Dashboard extends React.Component {
                                 </ButtonGroup>
                             </Grid>
                             <Grid item>
-                                <AddMealButton label="+ Add meal" />
+                                <AddMealDialog label="+ Add meal" />
                             </Grid>
                         </Grid>
